match(gruntcombat): ComputeFacing 88.3->100% - unsigned m_timePerTile->double
Retail zero-extends m_timePerTile ([ecx+0x41c]) to 64-bit before fild (mov edx,fld;
[esp]=edx; [esp+4]=0; fild qword) = (double)(unsigned), but the source cast the
signed i32 member giving fild dword. Cast to u32 locally (member stays i32 - it is
halved for kind 0x37 where signedness may matter). Byte-exact.

+++ src/Gruntz/GruntCombat.cpp
@@ -277,7 +277,8 @@ void CGrunt::ComputeFacing(double dt) {
     CWwdGameObjectA* h = m_10;
     double dx = static_cast<double>(m_lastTilePxX) - static_cast<double>(h->m_screenX);
     double dy = static_cast<double>(m_lastTilePxY) - static_cast<double>(h->m_screenY);
-    m_400 = (sqrt(dx * dx + dy * dy) / static_cast<double>(m_timePerTile)) * dt;
+    // retail zero-extends m_timePerTile to 64-bit before fild (unsigned->double)
+    m_400 = (sqrt(dx * dx + dy * dy) / static_cast<double>(static_cast<u32>(m_timePerTile))) * dt;
     m_408 = static_cast<double>(h->m_screenX);
     m_410 = static_cast<double>(h->m_screenY);
 }
